refactor(codegen): extract bus publish + key-filter into bus/dispatch.rs (Round 3b)
3 fns lifted into BusDispatch<'ctx>: lower_send (the heart of the
publish path — handles unicast, multicast, routing-key dispatch
including the fail-policy plumbing), lower_send_shm_ring (Form K
transport variant), lower_subscribe_key_filter (the runtime guard
keyed subscribers run against incoming wire subjects).

Surface bumps to pub(crate):
- Cx fields: bus_state, current_self, shm_ring_subjects,
  routing_key_subjects, current_user_fn_fallible,
  current_user_fn_exit_bb
- Cx methods: mangled_for_path, key_value_to_i64_pair, lower_stmt,
  llvm_basic_type, populate_user_type_fields,
  build_alloc_bus_unmatched_key
- Structs: BusState, RoutingKeySubjectInfo (+ fields),
  ShmRingBindingInfo, SelfCx (+ fields), FallibleCtx (+ fields),
  Scope (+ locals field), BlockEnd

codegen.rs: 36143 -> 35350 (-793 lines).
