+++ crates/hale-codegen/src/bus/mod.rs
@@ -1,3 +1,4 @@
 //! Bus codegen (publish, subscribe, wire, runtime). Round 3 of the codegen model-org refactor.
 
+pub(crate) mod dispatch;
 pub(crate) mod wire;
